+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * `FluxContainer`: Uses functional version of `setState`
 * `FluxMixin`: Subscriptions are setup in `componentWillMount` rather than
  `componentDidMount`
-* `React` added as dependency to `flux/utils`
+* `React` added as peer dependency to `flux/utils`
 * Package `dist/FluxUtils.js` alongside `dist/Flux.js`
 
 _**Note**: This is marked as a breaking change due to the large number of small