+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * `FluxContainer`: Uses functional version of `setState`
 * `FluxMixin`: Subscriptions are setup in `componentWillMount` rather than
  `componentDidMount`
+* `Dispatcher`: Methods `register` and `unregister` can not be called in the
+ middle of a dispatch
 * `React` added as peer dependency to `flux/utils`
 * Package `dist/FluxUtils.js` alongside `dist/Flux.js`